// src/common/interfaces/api-response.interface.ts
export interface ApiResponse<T = any> {
  success: boolean;
  message: string;
  data?: T;
  meta?: {
    page?: number;
    limit?: number;
    total?: number;
    totalPages?: number;
  };
  error?: {
    code: string;
    details?: Record<string, any>;
  };
  timestamp?: string;
  statusCode: number 
}